+++ src/jupyter_black/jupyter_black.py
@@ -65,7 +65,23 @@ class BlackFormatter:
         self.mode = mode
 
         self.is_lab = is_lab
-        if not is_lab:
+        if is_lab:
+            js_func = """
+                <script type="application/javascript" id="jupyter_black">
+                (function() {
+                    if (window.IPython === undefined) {
+                        return
+                    }
+                    var msg = "WARNING: it looks like you might have loaded " +
+                        "jupyter_black in a non-lab notebook with " +
+                        "`is_lab=True`. Please double check, and if " +
+                        "loading with `%load_ext` please review the README!"
+                    console.log(msg)
+                    alert(msg)
+                })()
+                </script>
+                """
+        else:
             js_func = """
                 <script type="application/javascript" id="jupyter_black">
                 function jb_set_cell(
@@ -80,11 +96,11 @@ class BlackFormatter:
                 }
                 </script>
                 """
-            display(
-                HTML(js_func),
-                display_id="jupyter_black",
-                update=False,
-            )
+        display(
+            HTML(js_func),
+            display_id="jupyter_black",
+            update=False,
+        )
 
     @staticmethod
     def _config_from_pyproject_toml() -> t.Dict[str, t.Any]: