+++ tests/test_lab.py
@@ -32,13 +32,10 @@ def test_lab(lab: t.Callable) -> None:
 
 
 def test_empty_cell(lab: t.Callable) -> None:
-    """Empty cells shouldn't break thinks."""
+    """Empty cells shouldn't break things."""
     cells = [
         {
-            "source": ["import jupyter_black"],
-        },
-        {
-            "source": ["jupyter_black.load(line_length=79, lab=True)"],
+            "source": ["%load_ext jupyter_black"],
         },
         {
             "source": [],
@@ -60,13 +57,8 @@ def test_empty_cell(lab: t.Callable) -> None:
     assert fix_quotes_with_magic[-1] == 'print("foo")'
 
 
-def test_lab_loadext_fails(lab: t.Callable) -> None:
-    """Fail with `%load_ext jupyter_black`.
-
-    Currently the default is `jupyter_black.load(lab=False)`. When loading by
-    `%load_ext`, one cannot specify any configuration, so it should fail to
-    blacken the cells.
-    """
+def test_lab_loadext(lab: t.Callable) -> None:
+    """Test loading with %load_ext magic."""
     cells = [
         {
             "source": ["%load_ext jupyter_black"],
@@ -82,7 +74,7 @@ def test_lab_loadext_fails(lab: t.Callable) -> None:
     ]
     output = lab(cells)
     fix_quotes = source_from_cell(output, "singlequotes")
-    assert fix_quotes[-1] == "print('foo')"
+    assert fix_quotes[-1] == 'print("foo")'
 
     fix_quotes_with_magic = source_from_cell(output, "magic_singlequotes")
-    assert fix_quotes_with_magic[-1] == "print('foo')"
+    assert fix_quotes_with_magic[-1] == 'print("foo")'